+++ mat.h
@@ -4,7 +4,7 @@
 
 typedef struct
 {
-	float *data;
+	double *data;
 	int *shape;
 	int dim;
 	int size;
@@ -26,7 +26,7 @@ mat_t *
 mat_alloc(int dim, int *shape);
 
 mat_t *
-mat_init(int dim, int *shape, float v);
+mat_init(int dim, int *shape, double v);
 
 mat_t *
-vec_init(int size, float v);
+vec_init(int size, double v);
